adding Lebesgue measure
